Updated vertex group export
Added "Clean Vertex Groups" export option. This always happened before, but now it can be disabled.
Exporter now checks for and warns about local bones that will be deleted when loaded by the game.
Exporter now checks for and warns about vertices that are in more than four vertex groups.
Fixed export error when object was active but not selected.